Refactoring: actions/search
